+++ .gitignore
@@ -1,1 +1,2 @@
 images/.DS_Store
+node_modules
