+++ todoList.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>ToDo List</title>
     <link rel="stylesheet" href="todoList.css" />
-    <script type="module" src="todoList.js"></script>
   </head>
   <body>
     <div class="container">
@@ -68,5 +67,7 @@
         </div>
       </div>
     </div>
+
+    <script type="module" src="todoList.js"></script>
   </body>
 </html>
